move format time to time util
