+++ src/utils/logging/index.ts
@@ -1,6 +1,6 @@
 import { ElectronFile } from 'types/upload';
 import { convertBytesToHumanReadable } from 'utils/file';
-import { formatDateTime } from 'utils/file';
+import { formatDateTime } from 'utils/time';
 import { saveLogLine, getLogs } from 'utils/storage';
 
 export function addLogLine(log: string) {
